moved modal to home and set up audio API call
