+++ client/src/pages/AboutUs.js
@@ -42,19 +42,6 @@ const programmers = [
 ]
 
 const AboutUs = () => {
-  const [show,setShow] = useState(false)
-
-  const handleClose = () => {
-    setShow(false)
-  };
-
-  const handleShow =()=>{
-    setShow(true)
-  };
-
-  const handleSave = () => {
-    setShow(false)
-  };
 
   const renderProgrammers = () =>{
     return programmers.map((developer)=>{
@@ -86,7 +73,6 @@ const AboutUs = () => {
                   style = {{height: "2rem"}}
                   />
                 </Nav.Link>
-                {/* <Nav.Link href = {person.LinkedIn}>Personal Website</Nav.Link> */}
               </Nav.Item>
             </Nav>
           </Card.Footer>
@@ -98,8 +84,6 @@ const AboutUs = () => {
   return (
     <Container>
       <h1>About Us</h1>
-      <EntryModal handleClose = {handleClose} handleSave = {handleSave} show = {show} />
-      <button onClick = {handleShow}>ShowModal</button>
 
       <Row md = {1} lg = {2}>
         {renderProgrammers()}
